requirements.txt
Flask==2.3.3
requests==2.31.0
debugpy # Required for debugging.
gradio>=6.0.0
google-genai>=0.1.0
googlemaps
